Refactored Title code; Refactored Style object;
- src/main/kotlin/main.kt
  - Removed code pertaining to the Title Text.

- src/main/kotlin/screens/edit/EditFile.kt
  - Modified code for compabitility with changes to the Custom Style
    object.

- src/main/kotlin/screens/title/Title.kt
  - Added new file to hold Title Text code.

- src/main/kotlin/screens/welcome/Welcome.kt
  - Modified code for compatibility with changes to the Custom Style
    object.

- src/main/kotlin/util/CustomStyle.kt
  - Moved specified numbers into Objects.
