+++ .idea/codeStyles/Project.xml
@@ -18,6 +18,9 @@
       <option name="NAME_COUNT_TO_USE_STAR_IMPORT_FOR_MEMBERS" value="2147483647" />
       <option name="CODE_STYLE_DEFAULTS" value="KOTLIN_OFFICIAL" />
     </JetCodeStyleSettings>
+    <Properties>
+      <option name="KEEP_BLANK_LINES" value="true" />
+    </Properties>
     <Shell_Script>
       <option name="SWITCH_CASES_INDENTED" value="true" />
       <option name="REDIRECT_FOLLOWED_BY_SPACE" value="true" />
