+++ src/main/kotlin/screens/title/Title.kt
@@ -23,39 +23,28 @@
  *
  */
 
-package util
+package screens.title
 
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import util.CustomStyle
+import util.Resource
 
-object CustomStyle {
-    private object FontSize {
-        val titleText = 24.sp
-        val paragraphText = 16.sp
-    }
+@Composable
+fun title() {
+    val headerText = Resource.getStringResource("application-header")
 
-    object PadValue {
-        val standard = 16.dp;
-        val double = standard * 2;
-    }
-
-
-    fun titleText(): TextStyle {
-        return TextStyle(
-            color = Color.Black,
-            fontWeight = FontWeight.ExtraBold,
-            fontSize = FontSize.titleText
-        )
-    }
-
-    fun paragraphText(): TextStyle {
-        return TextStyle(
-            color = Color.Black,
-            fontWeight = FontWeight.Normal,
-            fontSize = FontSize.paragraphText
+    // Title Row
+    Row(
+        horizontalArrangement = Arrangement.Center,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = headerText,
+            style = CustomStyle.titleText()
         )
     }
 }
